Phase6 検分指摘修正: 活動記録全面再構築・CMS/ダイジェスト/ギャラリー修正
F06.4 活動記録: EAV→JSON field_values方式に全面再構築、11エンドポイント追加、URL修正
F06.3 ダイジェスト: @Async自己呼び出し修正(DigestAsyncExecutor抽出)、DTO補完
F06.1 CMS: 6エンドポイント追加(auto-save/bulk/RSS/share/self-review)、FK制約追加
F06.2 ギャラリー: クエリパラメータ追加(from/to/visibility/photoIds)、fileSize型修正

+++ backend/src/main/java/com/mannschaft/app/gallery/dto/PhotoResponse.java
@@ -18,7 +18,7 @@ public class PhotoResponse {
     private final String thumbnailS3Key;
     private final String originalFilename;
     private final String contentType;
-    private final Integer fileSize;
+    private final Long fileSize;
     private final Integer width;
     private final Integer height;
     private final String caption;
